# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intelligent Data Dictionary Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Dataset {
  id: number;
  name: string;
  rowCount: number;
  columnCount: number;
  fileSize: number;
  /** uploading | analyzing | ready | error */
  status: string;
  /** @nullable */
  aiSummary?: string | null;
  createdAt: string;
}

export interface ColumnInfo {
  id: number;
  datasetId: number;
  name: string;
  /** string | integer | float | boolean | date | email | currency | percentage | category */
  dataType: string;
  nullable?: boolean;
  nullCount: number;
  nullPercent: number;
  uniqueCount: number;
  uniquePercent: number;
  sampleValues: string[];
  /** @nullable */
  minValue?: string | null;
  /** @nullable */
  maxValue?: string | null;
  isPrimaryKey?: boolean;
  isForeignKey?: boolean;
  /** @nullable */
  foreignKeyRef?: string | null;
  /** @nullable */
  aiDescription?: string | null;
  /** @nullable */
  businessMeaning?: string | null;
  /** @nullable */
  suggestedUsage?: string | null;
  /** @nullable */
  dataQualityNotes?: string | null;
  semanticTags?: string[];
}

export interface Relationship {
  id: number;
  datasetId: number;
  fromColumn: string;
  toColumn: string;
  /** primary_key | foreign_key | categorical | reference */
  relationshipType: string;
  /** @nullable */
  description?: string | null;
}

export interface DatasetDetail {
  id: number;
  name: string;
  rowCount: number;
  columnCount: number;
  fileSize: number;
  status: string;
  /** @nullable */
  aiSummary?: string | null;
  /** @nullable */
  businessGlossary?: string | null;
  /** @nullable */
  dataQualityScore?: number | null;
  duplicateRows?: number;
  createdAt: string;
  columns: ColumnInfo[];
  relationships: Relationship[];
}

export interface AnalysisResult {
  datasetId: number;
  status: string;
  /** @nullable */
  aiSummary?: string | null;
  /** @nullable */
  businessGlossary?: string | null;
  /** @nullable */
  dataQualityScore?: number | null;
}

export type DatasetStatsColumnTypeBreakdownItem = {
  type: string;
  count: number;
};

export type DatasetStatsNullDistributionItem = {
  column: string;
  nullPercent: number;
};

export interface DatasetStats {
  totalColumns: number;
  numericFields: number;
  categoricalFields: number;
  dateFields: number;
  stringFields: number;
  missingValueCount: number;
  duplicateRows: number;
  relationshipCount: number;
  dataQualityScore: number;
  nullableColumns: number;
  primaryKeyCandidates: number;
  columnTypeBreakdown?: DatasetStatsColumnTypeBreakdownItem[];
  nullDistribution?: DatasetStatsNullDistributionItem[];
}

export interface ErrorResponse {
  error: string;
}