# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intelligent Data Dictionary Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisResult {
  datasetId: number;
  status: string;
  /** @nullable */
  aiSummary?: string | null;
  /** @nullable */
  businessGlossary?: string | null;
  /** @nullable */
  dataQualityScore?: number | null;
}