# lib/api-zod/src/generated/types/columnInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intelligent Data Dictionary Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ColumnInfo {
  id: number;
  datasetId: number;
  name: string;
  /** string | integer | float | boolean | date | email | currency | percentage | category */
  dataType: string;
  nullable?: boolean;
  nullCount: number;
  nullPercent: number;
  uniqueCount: number;
  uniquePercent: number;
  sampleValues: string[];
  /** @nullable */
  minValue?: string | null;
  /** @nullable */
  maxValue?: string | null;
  isPrimaryKey?: boolean;
  isForeignKey?: boolean;
  /** @nullable */
  foreignKeyRef?: string | null;
  /** @nullable */
  aiDescription?: string | null;
  /** @nullable */
  businessMeaning?: string | null;
  /** @nullable */
  suggestedUsage?: string | null;
  /** @nullable */
  dataQualityNotes?: string | null;
  semanticTags?: string[];
}