# lib/api-zod/src/generated/types/dataset.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intelligent Data Dictionary Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface Dataset {
  id: number;
  name: string;
  rowCount: number;
  columnCount: number;
  fileSize: number;
  /** uploading | analyzing | ready | error */
  status: string;
  /** @nullable */
  aiSummary?: string | null;
  createdAt: Date;
}