# lib/api-zod/src/generated/types/datasetDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intelligent Data Dictionary Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { ColumnInfo } from './columnInfo';
import type { Relationship } from './relationship';

export interface DatasetDetail {
  id: number;
  name: string;
  rowCount: number;
  columnCount: number;
  fileSize: number;
  status: string;
  /** @nullable */
  aiSummary?: string | null;
  /** @nullable */
  businessGlossary?: string | null;
  /** @nullable */
  dataQualityScore?: number | null;
  duplicateRows?: number;
  createdAt: Date;
  columns: ColumnInfo[];
  relationships: Relationship[];
}