# lib/api-zod/src/generated/types/datasetStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intelligent Data Dictionary Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { DatasetStatsColumnTypeBreakdownItem } from './datasetStatsColumnTypeBreakdownItem';
import type { DatasetStatsNullDistributionItem } from './datasetStatsNullDistributionItem';

export interface DatasetStats {
  totalColumns: number;
  numericFields: number;
  categoricalFields: number;
  dateFields: number;
  stringFields: number;
  missingValueCount: number;
  duplicateRows: number;
  relationshipCount: number;
  dataQualityScore: number;
  nullableColumns: number;
  primaryKeyCandidates: number;
  columnTypeBreakdown?: DatasetStatsColumnTypeBreakdownItem[];
  nullDistribution?: DatasetStatsNullDistributionItem[];
}