# lib/api-zod/src/generated/types/relationship.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intelligent Data Dictionary Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface Relationship {
  id: number;
  datasetId: number;
  fromColumn: string;
  toColumn: string;
  /** primary_key | foreign_key | categorical | reference */
  relationshipType: string;
  /** @nullable */
  description?: string | null;
}